requirements.txt
### Main package ###
Django
Pillow

### RESTful ###
djangorestframework

### Tool ###
gunicorn

### Database ###
psycopg2-binary

### Filters ###
django-filter

### CORS ###
django-cors-headers

### Swagger ###
drf-yasg